updated lora positions
